requirements.txt
fastapi==0.110.0
mysql-connector-python==8.3.0
pandas==2.2.1
